+++ packages/brick-icons/src/BrickIcon.tsx
@@ -1,33 +1,47 @@
 import React from "react";
-import categories from "./categories";
+import { hasOwnProperty } from "@next-core/brick-utils";
+import categories from "./generated/categories";
 
 export interface BrickIconProps {
   icon: string;
   category?: string;
 }
 
-function isValidKey(key: string, obj: {}): key is keyof typeof obj {
-  return key in obj;
-}
+export const BrickIcon = ({
+  category,
+  icon,
+}: BrickIconProps): React.ReactElement => {
+  const actualCategory = category || "default";
 
-export const BrickIcon = (props: BrickIconProps): React.ReactElement => {
-  const category = props.category || "default";
+  const [iconComponent, setIconComponent] = React.useState<SvgrComponent>();
 
-  const iconProps: Record<string, any> = {
-    // Reset width and height.
-    width: null,
-    height: null,
-    className: `easyops-icon easyops-icon-${category || "default"}-${
-      props.icon
-    }`
-  };
+  React.useEffect(() => {
+    (async () => {
+      let component: SvgrComponent;
+      // Avoid prototype collisions, such as `actualCategory: "toString"`.
+      if (hasOwnProperty(categories, actualCategory)) {
+        let categoryIcons: Record<string, SvgrComponent> = {};
+        try {
+          categoryIcons = await categories[actualCategory]();
+        } catch (error) {
+          // eslint-disable-next-line no-console
+          console.error("Load icons failed:", error);
+        }
+        // Avoid prototype collisions, such as `icon: "toString"`.
+        if (hasOwnProperty(categoryIcons, icon)) {
+          component = categoryIcons[icon];
+        }
+      }
+      setIconComponent(() => component);
+    })();
+  }, [icon, actualCategory]);
 
-  const categoryIcons: Record<string, SvgrComponent> =
-    isValidKey(category, categories) && categories[category];
-
-  if (categoryIcons && categoryIcons[props.icon]) {
-    return React.createElement(categoryIcons[props.icon], iconProps);
-  }
-
-  return null;
+  return iconComponent
+    ? React.createElement(iconComponent, {
+        // Reset width and height.
+        width: null,
+        height: null,
+        className: `easyops-icon easyops-icon-${actualCategory}-${icon}`,
+      })
+    : null;
 };