+++ packages/brick-kit/src/BrickAsComponent.spec.tsx
@@ -7,6 +7,11 @@ import {
   ForwardRefSingleBrickAsComponent,
 } from "./BrickAsComponent";
 import * as runtime from "./core/Runtime";
+import * as transformProperties from "./transformProperties";
+import {
+  RuntimeBrickElementWithTplSymbols,
+  symbolForParentRefForUseBrickInPortal,
+} from "./core/exports";
 
 const bindListeners = jest.spyOn(listenerUtils, "bindListeners");
 const spyOnResolve = jest.fn(
@@ -17,6 +22,10 @@ const spyOnResolve = jest.fn(
 const _internalApiGetRouterState = jest
   .spyOn(runtime, "_internalApiGetRouterState")
   .mockReturnValue("mounted");
+const sypOnTransformProperties = jest.spyOn(
+  transformProperties,
+  "transformProperties"
+);
 jest.spyOn(runtime, "_internalApiGetResolver").mockReturnValue({
   resolve: spyOnResolve,
 } as any);
@@ -41,6 +50,7 @@ describe("BrickAsComponent", () => {
   });
 
   it("should work", async () => {
+    const mockRef = {} as React.RefObject<HTMLElement>;
     const wrapper = mount(
       <BrickAsComponent
         useBrick={{
@@ -57,6 +67,7 @@ describe("BrickAsComponent", () => {
         data={{
           tips: "good",
         }}
+        parentRefForUseBrickInPortal={mockRef}
       />
     );
 
@@ -70,6 +81,11 @@ describe("BrickAsComponent", () => {
       },
     });
     expect((div as RuntimeBrickElement).$$typeof).toBe("native");
+    expect(
+      (div as RuntimeBrickElementWithTplSymbols)[
+        symbolForParentRefForUseBrickInPortal
+      ]
+    ).toBe(mockRef);
   });
 
   it("should work for multiple bricks", async () => {
@@ -107,6 +123,7 @@ describe("BrickAsComponent", () => {
   });
 
   it("should work for `if`", async () => {
+    sypOnTransformProperties.mockClear();
     const wrapper = mount(
       <BrickAsComponent
         useBrick={[
@@ -135,6 +152,7 @@ describe("BrickAsComponent", () => {
     const span = wrapper.find("span").getDOMNode() as HTMLDivElement;
     expect(span.title).toBe("better");
     expect(wrapper.find("div").length).toBe(0);
+    expect(sypOnTransformProperties).toBeCalledTimes(1);
   });
 
   it("should work for unsupported `resolvable-if`", async () => {