+++ packages/dev-dependencies/patches/index.js
@@ -9,3 +9,5 @@ exports.addPostBuildScriptForLibs = require("./addPostBuildScriptForLibs");
 exports.disableSdkRenovate = require("./disableSdkRenovate");
 exports.migrateOfficialRenovate = require("./migrateOfficialRenovate");
 exports.migrateHusky = require("./migrateHusky");
+exports.addMocksForBrickIcons = require("./addMocksForBrickIcons");
+exports.migrateJest = require("./migrateJest");